[DI] Improve separation between RC and breakpoint logic (#4992)
This will make it easier to mock either one or the other in tests or to
add probes without relying on RC.
